添加分类DAO

+++ JBlog-JPA/src/main/java/com/youthlin/jblog/constant/Constant.java
@@ -12,8 +12,8 @@ public class Constant {
     public static final String POST_TYPE_IMAGE = "image";
     public static final String CATEGORY_TEXT_UNCATEGORY_NAME = "默认分类";
     public static final String CATEGORY_IMAGE_UNCATEGORY_NAME = "默认相册";
-    public static final int CATEGORY_TEXT_TYPE = 0;
-    public static final int CATEGORY_IMAGE_TYPE = 1;
+    public static final byte CATEGORY_TEXT_TYPE = 0;
+    public static final byte CATEGORY_IMAGE_TYPE = 1;
 
     public static final String SETTINGS_TEXT_COUNT_PER_PAGE = "text_count_per_page";
     public static final String SETTINGS_IMAGE_COUNT_PER_PAGE = "image_count_per_page";
